[#2612]Move class.
