+++ GAE/src/org/akvo/flow/api/export/SurveyRestResponse.java
@@ -14,7 +14,7 @@
  *  The full license text can also be seen at <http://www.gnu.org/licenses/agpl.html>.
  */
 
-package org.waterforpeople.mapping.app.web.dto;
+package org.akvo.flow.api.export;
 
 import java.util.List;
